Split generator and subscription tests into 7 file
