+++ volunteer/tests/__init__.py
@@ -7,5 +7,10 @@ from . import test_volunteer_generator_subscription_common
 
 from . import test_volunteer_shift
 from . import test_volunteer_shift_participation
-from . import test_volunteer_shift_recurrent_subscription
-from . import test_volunteer_shift_recurrent_generator
+from . import test_volunteer_shift_recurrent_generator_constraints
+from . import test_volunteer_shift_recurrent_generator_permissions
+from . import test_volunteer_shift_recurrent_generator_shift_generation
+from . import test_volunteer_shift_recurrent_subscription_capacity
+from . import test_volunteer_shift_recurrent_subscription_participation
+from . import test_volunteer_shift_recurrent_subscription_permissions
+from . import test_volunteer_shift_recurrent_subscription_uniqueness
